backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
websockets==13.1
python-dotenv==1.0.1
ollama==0.4.4
sqlalchemy==2.0.36

